clean up and rename classes for javax UpgradeRequest interface
